Created Database Models and Migration Support
Created the following database models: Student for storing information
about the students, Parent for storing information about the parents,
Tutor for storing information about the peripatetic tutors, Lesson for
storing information about the peripatetic lessons, Room for storing
information about the rooms and Instrument for storing information about
the school instruments. StudentLessonAssociation and
LessonTutorAssociation are used to create the many to many relationship
between Student and Lesson as well as Lesson and Tutor respectively.

Flask-Migrate was also added to allow easy upgrading of the database
with any future changes. The Migrate command has be added to the Manager
object and is available via manage.py.