Created Site And Login System.
Created Home and Auth blueprints. The Home blueprint acts as the
application 'homepage' which will contain information about the
application/lessons being run by the school as well as the terms and
conditions etc.

The auth blueprint will contain all the views/forms for authenticating
users to the application. I have currently implemented login_manager,
but not completed views for register/authenticating users to the
application yet. I have implemented forms and form validation on the
register and login forms.

To complete the login system, the login and register views need to be
completed.

Static files and templates, including the __layout__.html template,
which acts as the master layout template have been added.