Updated Login System and Implemented Roles
Updated the login system, to be aware of user roles, and redirect a user
to their dashboard based on these roles. Create a custom login_required
decorator, do decorate views based on roles.

Added a utils module to the auth blueprint for the login_required
decorator, more will be added to this module as the application is
developed.

Added a student blueprint along with a dashboard, navbar and pills
template. This creates a very simple dashboard which will become the
main screen seen by students using the application.

Added a staff and tutor blueprint, both with 'test' dashboard views,
which only return a plain text response for testing the login_required
decorator with their roles.

Made sure all code was formatted according to convention, all imports,
even from the same module, on separate lines.