Updated Custom Login Decorator
Updated custom login_required decorator so that id does not call
login_manager.unauthorized() when a role is not met. Instead it calls a
secondary function in the auth.utils module which displays an error
message to the user and redirects them back to their dashboard.

Login form templates have also been updated to display 'Password'
placeholder in password input boxes.