Implemented Student Blueprint
I have implemented the student blueprint, and all of the views and
templates required for this blueprint. Students are now able to view
their lessons, tutors, staff and attendance.

These views allow students to view all lessons/tutors/staff as well as
individual records for each.

This blueprint will be the only blueprint students can access, except
the auth blueprint.