Added Flask-Admin
Added Flask-Admin for use by staff only, and this is protected by roles,
to edit database entries if (and hopefully not when) needed...

Useful for development and testing, but also provides an easy way to see
all the data stored in one place. Will, in future, provide 'clean up'
facilities, where all data from the previous year will be delete.